Refactor token provider logic to shared extension method for 2FA tokens
